% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chute.R
\name{chute}
\alias{chute}
\title{"Chute" a ladder into a file}
\usage{
chute(ladder, file)
}
\arguments{
\item{ladder}{From \code{\link{ladder}}}

\item{file}{Yknow, a filename. png pls}
}
\description{
"Chute" a ladder into a file
}
